backend/requirements.txt
# Core utilities
tqdm
pillow
rapidfuzz
requests
python-dotenv

# CLIP & embeddings (for future RAG implementation)
open-clip-torch
ftfy
regex
numpy

# FastAPI backend
fastapi
uvicorn[standard]
python-multipart
httpx  # Async HTTP client for API calls

# Google Gemini SDK
google-genai

# Note: Install PyTorch + FAISS via conda (see environment.yml)
# For pip-only installs:
#   pip install torch torchvision --index-url https://download.pytorch.org/whl/cu121
#   pip install faiss-cpu
